Feat: filterHandler
